% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_JSON_helpers.R
\name{convert_Stan_data_list_to_JSON}
\alias{convert_Stan_data_list_to_JSON}
\title{convert_Stan_data_list_to_JSON}
\usage{
convert_Stan_data_list_to_JSON(Stan_data_list, data_dir = NULL)
}
\description{
convert_Stan_data_list_to_JSON
}
\keyword{internal}
