% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_create_summary_and_traces.R
\name{create_summary_and_traces}
\alias{create_summary_and_traces}
\title{create_summary_and_traces}
\usage{
create_summary_and_traces(
  model_results,
  init_object,
  n_nuisance,
  compute_main_params = TRUE,
  compute_transformed_parameters = TRUE,
  compute_generated_quantities = TRUE,
  save_log_lik_trace = FALSE,
  save_nuisance_trace = FALSE,
  compute_nested_rhat = FALSE,
  n_superchains = NULL,
  save_trace_tibbles = FALSE
)
}
\description{
create_summary_and_traces
}
\keyword{internal}
