% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_create_superchain_ids.R
\name{create_superchain_ids}
\alias{create_superchain_ids}
\title{#' create_superchain_ids}
\usage{
create_superchain_ids(n_superchains, n_chains)
}
\description{
#' create_superchain_ids
}
\keyword{internal}
