% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_init_burnin.R
\name{init_and_run_burnin}
\alias{init_and_run_burnin}
\title{init_and_run_burnin}
\usage{
init_and_run_burnin(
  Model_type,
  init_object,
  Stan_data_list,
  y,
  N,
  parallel_method,
  manual_tau,
  tau_if_manual,
  n_chains_burnin,
  sample_nuisance,
  Phi_type,
  n_params_main,
  n_nuisance,
  seed,
  n_burnin,
  adapt_delta,
  LR_main,
  LR_us,
  n_adapt,
  partitioned_HMC,
  diffusion_HMC,
  clip_iter,
  gap,
  metric_type_main,
  metric_shape_main,
  metric_type_nuisance,
  metric_shape_nuisance,
  max_eps_main,
  max_eps_us,
  max_L,
  tau_mult,
  ratio_M_us,
  ratio_M_main,
  interval_width_main,
  interval_width_nuisance,
  force_autodiff,
  force_PartialLog,
  multi_attempts,
  n_nuisance_to_track,
  Model_args_as_Rcpp_List
)
}
\description{
init_and_run_burnin
}
\keyword{internal}
