% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_bs_model.R
\name{init_bs_model_external}
\alias{init_bs_model_external}
\title{init_bs_model_external}
\usage{
init_bs_model_external(Stan_data_list, Stan_model_file_path)
}
\description{
init_bs_model_external
}
\keyword{internal}
