% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_init_burnin.R
\name{init_EHMC_args_as_Rcpp_List}
\alias{init_EHMC_args_as_Rcpp_List}
\title{init_EHMC_args_as_Rcpp_List}
\usage{
init_EHMC_args_as_Rcpp_List(diffusion_HMC)
}
\description{
init_EHMC_args_as_Rcpp_List
}
\keyword{internal}
