% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_init_hard_coded_models.R
\name{init_hard_coded_model}
\alias{init_hard_coded_model}
\title{init_hard_coded_model}
\usage{
init_hard_coded_model(Model_type, y, N = N, model_args_list = list(), ...)
}
\description{
init_hard_coded_model
}
\keyword{internal}
