% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_init_inits.R
\name{init_inits}
\alias{init_inits}
\title{init_inits}
\usage{
init_inits(
  init_model_outs,
  init_lists_per_chain,
  compile,
  force_recompile,
  cmdstanr_model_fit_obj,
  n_chains_burnin,
  n_params_main,
  n_nuisance,
  N,
  sample_nuisance,
  Stan_model_file_path,
  Stan_data_list,
  Stan_cpp_user_header,
  Stan_cpp_flags,
  ...
)
}
\description{
init_inits
}
\keyword{internal}
