% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_init_model.R
\name{init_model}
\alias{init_model}
\title{init_model}
\usage{
init_model(
  Model_type,
  y,
  N,
  n_params_main = NULL,
  n_nuisance = NULL,
  X = NULL,
  model_args_list,
  Stan_model_file_path,
  Stan_cpp_user_header,
  Stan_data_list,
  ...
)
}
\description{
init_model
}
\keyword{internal}
