% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_model_and_vals.R
\name{initialise_model}
\alias{initialise_model}
\title{initialise_model}
\usage{
initialise_model(
  Model_type,
  compile,
  force_recompile,
  cmdstanr_model_fit_obj,
  y,
  N,
  n_params_main,
  n_nuisance,
  init_lists_per_chain,
  sample_nuisance,
  n_chains_burnin,
  model_args_list,
  Stan_data_list,
  Stan_model_file_path,
  Stan_cpp_user_header,
  Stan_cpp_flags,
  ...
)
}
\description{
initialise_model
}
\keyword{internal}
