% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_plot_batch_trace_dens.R
\name{plot_multiple_params_batched}
\alias{plot_multiple_params_batched}
\title{plot_multiple_params_batched}
\usage{
plot_multiple_params_batched(
  draws_array,
  param_prefixes,
  plot_type = c("density", "trace"),
  batch_size = 9
)
}
\description{
plot_multiple_params_batched
}
\keyword{internal}
