% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_post_burnin_prep_inits.R
\name{R_fn_avg_matrix_columns}
\alias{R_fn_avg_matrix_columns}
\title{R_fn_avg_matrix_columns}
\usage{
R_fn_avg_matrix_columns(matrix, n_chains_per_superchain)
}
\description{
R_fn_avg_matrix_columns
}
\keyword{internal}
