% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_post_burnin_prep_inits.R
\name{R_fn_grp_mat_cols_by_first_col_per_superchain_grp}
\alias{R_fn_grp_mat_cols_by_first_col_per_superchain_grp}
\title{R_fn_grp_mat_cols_by_first_col_per_superchain_grp}
\usage{
R_fn_grp_mat_cols_by_first_col_per_superchain_grp(
  matrix,
  n_chains_per_superchain
)
}
\description{
R_fn_grp_mat_cols_by_first_col_per_superchain_grp
}
\keyword{internal}
