% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_post_burnin_prep_inits.R
\name{R_fn_post_burnin_prep_for_sampling}
\alias{R_fn_post_burnin_prep_for_sampling}
\title{R_fn_post_burnin_prep_for_sampling}
\usage{
R_fn_post_burnin_prep_for_sampling(
  n_chains_sampling,
  n_superchains,
  n_params_main,
  n_nuisance,
  theta_main_vectors_all_chains_input_from_R,
  theta_us_vectors_all_chains_input_from_R
)
}
\description{
R_fn_post_burnin_prep_for_sampling
}
\keyword{internal}
