% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_create_summary_and_traces.R
\name{R_fn_remove_log_lik_from_array}
\alias{R_fn_remove_log_lik_from_array}
\title{R_fn_remove_log_lik_from_array}
\usage{
R_fn_remove_log_lik_from_array(arr)
}
\description{
R_fn_remove_log_lik_from_array
}
\keyword{internal}
