% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sample.R
\name{R_fn_sample_model}
\alias{R_fn_sample_model}
\title{R_fn_sample_model}
\usage{
R_fn_sample_model(
  Model_type,
  init_object,
  init_lists_per_chain,
  parallel_method,
  Stan_data_list,
  model_args_list,
  y,
  N,
  manual_tau,
  tau_if_manual,
  sample_nuisance,
  diffusion_HMC,
  partitioned_HMC,
  vect_type,
  Phi_type,
  inv_Phi_type,
  n_params_main,
  n_nuisance,
  n_chains_burnin,
  n_chains_sampling,
  n_superchains,
  seed,
  n_burnin,
  n_iter,
  adapt_delta,
  LR_main,
  LR_us,
  n_adapt,
  clip_iter,
  gap,
  metric_type_main,
  metric_shape_main,
  metric_type_nuisance,
  metric_shape_nuisance,
  max_eps_main,
  max_eps_us,
  max_L,
  tau_mult,
  ratio_M_us,
  ratio_M_main,
  interval_width_main,
  interval_width_nuisance,
  force_autodiff,
  force_PartialLog,
  multi_attempts,
  n_nuisance_to_track
)
}
\description{
R_fn_sample_model
}
\keyword{internal}
