% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_pkg_example_path_and_wd.R
\name{set_pkg_example_path_and_wd}
\alias{set_pkg_example_path_and_wd}
\title{set_pkg_example_path_and_wd}
\usage{
set_pkg_example_path_and_wd()
}
\description{
set_pkg_example_path_and_wd
}
