% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_cov_Welford.R
\name{update_cov_Welford}
\alias{update_cov_Welford}
\title{#' update_cov_Welford}
\usage{
update_cov_Welford(delta_old, delta_new, ii, cov_mat)
}
\description{
#' update_cov_Welford
}
\keyword{internal}
